move Scoreboard.java to service package
