+++ lib/src/main/java/live/match/api/LiveScoreboardApiImpl.java
@@ -2,6 +2,7 @@ package live.match.api;
 
 import live.match.service.Match;
 import live.match.service.MatchService;
+import live.match.service.Scoreboard;
 
 public class LiveScoreboardApiImpl implements LiveScoreboardApi {
     private final MatchService matchService;
