+++ lib/src/test/java/live/match/api/LiveScoreboardApiImplTest.java
@@ -1,6 +1,7 @@
 package live.match.api;
 
 import live.match.service.Match;
+import live.match.service.Scoreboard;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
